+++ Node/src/main/java/com/github/krextouch/rtds/node/model/MoveRequest.java
@@ -1,4 +1,4 @@
-package com.github.krextouch.rtds.node.service.model;
+package com.github.krextouch.rtds.node.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
