+++ Node/src/main/java/com/github/krextouch/rtds/node/repository/ClientRepository.java
@@ -1,6 +1,6 @@
 package com.github.krextouch.rtds.node.repository;
 
-import com.github.krextouch.rtds.node.service.trafficcontrol.Coordinate;
+import com.github.krextouch.rtds.node.trafficcontrol.Coordinate;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.List;
